# backend/clients/ts-sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Patron API",
    "description": "An open source Patreon alternative with lower fees designed for creators who publish ongoing sequential content like books, podcasts, and comics.",
    "contact": {
      "name": "Patron Team",
      "email": "support@patron.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.patron.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/api-keys": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "List API keys with cursor-based pagination and optional filtering",
        "description": "# Errors\nReturns error if API keys database query fails or connection issues occur",
        "operationId": "list_api_keys",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "Starting point UUID for paginated API key results",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of API keys to return per page (default: 50, max: 100)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "description": "Only show active or inactive API keys",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API keys list retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeysListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to list API keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during API keys listing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Create a new API key",
        "description": "# Errors\nReturns error if permissions are invalid, database error, or key generation fails",
        "operationId": "create_api_key",
        "requestBody": {
          "description": "API key creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "API key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid API key data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to create API keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during API key creation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/api-keys/{api_key_id}": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "Get a specific API key by ID",
        "description": "# Errors\nReturns error if API key not found, access denied, or database connection error",
        "operationId": "get_api_key",
        "parameters": [
          {
            "name": "api_key_id",
            "in": "path",
            "description": "UUID of the API key to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to view API keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied - API key belongs to different user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "API key not found or access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during API key retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "API Keys"
        ],
        "summary": "Update an API key",
        "description": "# Errors\nReturns error if API key not found, access denied, or database update error",
        "operationId": "update_api_key",
        "parameters": [
          {
            "name": "api_key_id",
            "in": "path",
            "description": "UUID of the API key to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Updated API key data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "API key updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to update API keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied - cannot update API key owned by different user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "API key not found for update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during API key update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "API Keys"
        ],
        "summary": "Delete an API key (hard delete for security)",
        "description": "# Errors\nReturns error if API key not found, access denied, or database deletion error",
        "operationId": "delete_api_key",
        "parameters": [
          {
            "name": "api_key_id",
            "in": "path",
            "description": "UUID of the API key to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "API key deleted successfully"
          },
          "401": {
            "description": "Authentication required to delete API keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied - cannot delete API key owned by different user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "API key not found for deletion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during API key deletion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiKeysListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ApiKeyResponse"
        },
        "description": "List of API keys for a user",
        "example": [
          {
            "createdAt": "2023-01-01T00:00:00Z",
            "expiresAt": "2024-01-01T00:00:00Z",
            "id": "d290f1ee-6c54-4b01-90e6-d701748f0851",
            "isActive": true,
            "keyPrefix": "pk_live_",
            "lastUsedAt": "2023-01-01T12:00:00Z",
            "name": "Production API Key",
            "permissions": [
              "read:posts",
              "write:posts"
            ],
            "updatedAt": "2023-01-01T12:00:00Z",
            "userId": "b2c3d4e5-6f78-9012-bcde-f12345678901"
          }
        ]
      },
...
      "CreateApiKeyRequest": {
        "type": "object",
        "description": "Request payload for creating a new API key",
        "required": [
          "name"
        ],
        "properties": {
          "expiresAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the API key should expire (optional)",
            "example": "2024-01-01T00:00:00Z"
          },
          "name": {
            "type": "string",
            "description": "Name identifier for the new API key",
            "example": "Production API Key"
          },
          "permissions": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Array of permissions to grant to this key",
            "example": [
              "read:posts",
              "write:posts"
            ]
          }
        },
        "example": {
          "expiresAt": "2024-01-01T00:00:00Z",
          "name": "Production API Key",
          "permissions": [
            "read:posts",
            "write:posts"
          ]
        }
      },
      "CreateApiKeyResponse": {
        "type": "object",
        "description": "Response when creating a new API key - includes the actual key value",
        "required": [
          "id",
          "name",
          "key",
          "keyPrefix",
          "permissions",
          "isActive"
        ],
        "properties": {
          "createdAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Timestamp of when the API key was just created",
            "example": "2023-01-01T00:00:00Z"
          },
          "expiresAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Key expiration date (when provided)",
            "example": "2024-01-01T00:00:00Z"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the newly created API key",
            "example": "a1b2c3d4-5e6f-7890-abcd-ef1234567890"
          },
          "isActive": {
            "type": "boolean",
            "description": "Initial active status of the new API key",
            "example": true
          },
          "key": {
            "type": "string",
            "description": "The actual API key value (only shown once during creation)",
            "example": "pk_live_1234567890abcdef"
          },
          "keyPrefix": {
            "type": "string",
            "description": "Prefix portion of the newly generated key",
            "example": "pk_live_"
          },
          "name": {
            "type": "string",
            "description": "Display name shown for this API key",
            "example": "Production API Key"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Permissions assigned to the new API key",
            "example": [
              "read:posts",
              "write:posts"
            ]
          }
        },
        "example": {
          "createdAt": "2023-01-01T00:00:00Z",
          "expiresAt": "2024-01-01T00:00:00Z",
          "id": "d290f1ee-6c54-4b01-90e6-d701748f0851",
          "isActive": true,
          "key": "pk_live_1234567890abcdef",
          "keyPrefix": "pk_live_",
          "name": "Production API Key",
          "permissions": [
            "read:posts",
            "write:posts"
          ]
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard JSON error response structure for API endpoints.",
        "required": [
          "error"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional error code for programmatic handling",
            "example": "AUTH_INVALID_CREDENTIALS"
          },
          "error": {
            "type": "string",
            "description": "Error message describing what went wrong",
            "example": "Invalid email or password"
          }
        },
        "example": {
          "code": "AUTH_INVALID_CREDENTIALS",
          "error": "Invalid email or password"
        }
      },
...
      "UpdateApiKeyRequest": {
        "type": "object",
        "description": "Request payload for updating an API key",
        "properties": {
          "expiresAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Update the expiration date",
            "example": "2024-06-01T00:00:00Z"
          },
          "isActive": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Change the active status of the key",
            "example": false
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "New name for the API key (optional)",
            "example": "Updated API Key Name"
          },
          "permissions": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Updated permissions array (optional)",
            "example": [
              "read:posts"
            ]
          }
        },
        "example": {
          "expiresAt": "2024-06-01T00:00:00Z",
          "isActive": false,
          "name": "Updated API Key Name",
          "permissions": [
            "read:posts"
          ]
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API key authentication using Bearer token in Authorization header"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid",
        "description": "Session-based authentication using secure HTTP-only cookies"
      }
    }
  },
...
}